Update README.md
Health Quality is important measure of accessing nation health. Health Quality can be measured 
on different scales, most important metrics being mortality- Infant, Child or Adult, Life 
Expectancy. Life Expectancy encompass mortality along the life cycle as compared to Adult or 
Child which accounts for a section of life cycle. 
Background studies indicate that Life Expectancy has improved a lot in the 20th centuries. Studies 
show different factors like advancements in Technology, Increase in Income levels and nations 
expenditure on Healthcare are major determinants of upsurge. Increased demand and need for 
Healthcare reachability led to spike in healthcare employment and improved health quality. 
Reforms in education furthered sanitation and personal health. However, deaths by different 
causes like diseases decreased the mean number of years in a person life. In this model we 
aggregated the above factors in available format to access Life Expectancy. 
Besides quality, the health expenditure also plays a key role in formulating health care policies 
and proper allocation of funds. 
Availability of health resources is one of key indicators influencing the health expenses of a 
country. Health Economics suggest that medical practices and equipment are the major 
determinants in the growth of health expenditure. Population age structure is often considered 
as a distinguisher of the health characteristics of a country. From the analysis of WHO health 
financing, the high-income countries have a positive effect of 0.477 on the expenses. The use of 
primary care gatekeepers seemed to result in lower health expenditure. Public sector provision 
of health services was associated with lower health expenditure